Returned generic map overlays logic to interactive map
