+++ src/app/shared/shared.module.ts
@@ -38,7 +38,6 @@ import { KeysPipe } from './keys.pipe';
 import { StationFlagComponent } from './station-flag/station-flag.component';
 import { ShakemapOverlaysPipe } from './shakemap-overlays.pipe';
 import { ShakemapMapOverlaysPipe } from './shakemap-map-overlays.pipe';
-import { MapOverlaysPipe } from './map-overlays.pipe';
 
 
 @NgModule({
@@ -81,8 +80,7 @@ import { MapOverlaysPipe } from './map-overlays.pipe';
     KeysPipe,
     StationFlagComponent,
     ShakemapOverlaysPipe,
-    ShakemapMapOverlaysPipe,
-    MapOverlaysPipe
+    ShakemapMapOverlaysPipe
   ],
   exports: [
     AttributionComponent,
